Remove base replay/recorder classes
