--- osu.Game/Screens/OnlinePlay/Matchmaking/RankedPlay/Cards/CardHandReplayPlayerBase.cs
@@ -1,47 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System;
-using osu.Framework.Graphics;
-using osu.Game.Online.RankedPlay;
-
-namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Cards
-{
-    public abstract partial class CardHandReplayPlayerBase : Component
-    {
-        /// <summary>
-        /// Maximum amount of frames that can get queued up at the same time
-        /// </summary>
-        public int MaxQueuedFrames { get; set; } = 20;
-
-        private int queuedFrames;
-        private double? lastPlayback;
-        private readonly OpponentCardHand cardHand;
-
-        protected CardHandReplayPlayerBase(OpponentCardHand cardHand)
-        {
-            this.cardHand = cardHand;
-        }
-
-        protected void EnqueueFrames(RankedPlayCardHandReplayFrame[] frames)
-        {
-            foreach (var frame in frames)
-            {
-                if (queuedFrames >= MaxQueuedFrames)
-                    return;
-
-                queuedFrames++;
-
-                double delay = Math.Max(lastPlayback != null ? lastPlayback.Value + frame.Delay - Time.Current : 0, 0);
-                lastPlayback = Time.Current + delay;
-
-                Scheduler.AddDelayed(() =>
-                {
-                    queuedFrames--;
-
-                    cardHand.SetState(frame.Cards);
-                }, delay);
-            }
-        }
-    }
-}
