+++ osu.Game/Screens/OnlinePlay/Matchmaking/RankedPlay/RankedPlayMatchInfo.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
@@ -74,6 +75,8 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay
         public bool IsOwnTurn => RoomState.ActiveUserId == client.LocalUser?.UserID;
 
         public int CurrentRound => RoomState.CurrentRound;
+
+        public int OpponentId => RoomState.Users.Keys.Single(u => u != client.LocalUser?.UserID);
 
         private readonly List<RankedPlayCardWithPlaylistItem> playerCards = new List<RankedPlayCardWithPlaylistItem>();
         private readonly List<RankedPlayCardWithPlaylistItem> opponentCards = new List<RankedPlayCardWithPlaylistItem>();
